Added editor view type tabs & refactor
