+++ src/lib/actions/EditorViewCode.svelte
@@ -1,0 +1,26 @@
+<script lang="ts">
+  import { createEventDispatcher } from "svelte";
+  import { editorViewType } from "$lib/stores.js";
+  import { iconButtonClasses } from "$lib/ui/iconButtonClasses.js";
+  import Tooltip from "$lib/ui/Tooltip.svelte";
+
+  const dispatch = createEventDispatcher();
+
+  function handleViewForm() {
+    $editorViewType = "code";
+    dispatch("viewForm");
+    document.dispatchEvent(new Event("viewFormButtonClick"));
+  }
+</script>
+
+<Tooltip label="Swap to code editor view.">
+  <button on:click={handleViewForm} class={iconButtonClasses($editorViewType === "code")}>
+    <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
+      <path
+        fill-rule="evenodd"
+        d="M12.316 3.051a1 1 0 01.633 1.265l-4 12a1 1 0 11-1.898-.632l4-12a1 1 0 011.265-.633zM5.707 6.293a1 1 0 010 1.414L3.414 10l2.293 2.293a1 1 0 11-1.414 1.414l-3-3a1 1 0 010-1.414l3-3a1 1 0 011.414 0zm8.586 0a1 1 0 011.414 0l3 3a1 1 0 010 1.414l-3 3a1 1 0 11-1.414-1.414L16.586 10l-2.293-2.293a1 1 0 010-1.414z"
+        clip-rule="evenodd"
+      />
+    </svg>
+  </button>
+</Tooltip>
