+++ src/lib/editor/Editor.svelte
@@ -1,341 +1,18 @@
 <script lang="ts">
-  import * as monaco from "monaco-editor";
-  import JSONWorker from "$lib/monaco-editor/json.worker.js?worker";
-  import EditorWorker from "monaco-editor/esm/vs/editor/editor.worker?worker";
-  import type { Environment } from "monaco-editor";
-  import type { Unsubscriber } from "svelte/store";
-
-  import { onDestroy, onMount } from "svelte";
-  import { slide } from "svelte/transition";
-  import {
-    editor,
-    editorProblems as problems,
-    goblError,
-    redoAvailable,
-    undoAvailable,
-    envelope,
-  } from "$lib/stores.js";
-  import EditorProblem from "./EditorProblem.svelte";
-  import WarningIcon from "$lib/ui/WarningIcon.svelte";
-  import ErrorIcon from "$lib/ui/ErrorIcon.svelte";
-  import SuccessIcon from "$lib/ui/SuccessIcon.svelte";
-  import LightbulbIcon from "$lib/ui/LightbulbIcon.svelte";
-
-  const modelUri = monaco.Uri.parse("gobl://doc.json");
+  import { editorViewType } from "$lib/stores.js";
+  import CodeEditor from "./code/CodeEditor.svelte";
+  import FormEditor from "./form/FormEditor.svelte";
 
   export let jsonSchemaURL: string;
-
-  let editorEl: HTMLElement;
-  let monacoEditor: monaco.editor.IStandaloneCodeEditor;
-  let model: monaco.editor.ITextModel;
-  let readOnlyEditHandler: monaco.IDisposable;
-  let lineNumber = 1;
-  let column = 1;
-  let drawerClosed = false;
-
-  let unsubscribeEditor: Unsubscriber;
-
-  // Sort by `monaco.MarkerSeverity` enum value descending, most severe shown first.
-  $: sortedProblems = $problems.sort((a, b) => b.severity - a.severity);
-  $: warningCount = $problems.filter((problem) => problem.severity === monaco.MarkerSeverity.Warning).length;
-  $: errorCount = $problems.filter((problem) => problem.severity === monaco.MarkerSeverity.Error).length;
-
-  $: {
-    setSchemaURI(jsonSchemaURL);
-  }
-
-  function setSchemaURI(uri: string) {
-    monaco.languages.json.jsonDefaults.setDiagnosticsOptions({
-      validate: true,
-      enableSchemaRequest: true,
-      schemaValidation: "warning",
-      schemas: uri
-        ? [
-            {
-              fileMatch: [modelUri.toString()],
-              uri,
-            },
-          ]
-        : [],
-    });
-    if (monacoEditor) {
-      const value = monacoEditor.getValue();
-      validateSchema(value);
-    }
-  }
-
-  onMount(() => {
-    (<Environment>(self as any).MonacoEnvironment) = {
-      getWorker: function (_: string, label: string) {
-        switch (label) {
-          case "json":
-            return new JSONWorker();
-          default:
-            return new EditorWorker();
-        }
-      },
-    };
-
-    model = monaco.editor.createModel("", "json", modelUri);
-
-    setSchemaURI(jsonSchemaURL);
-
-    monacoEditor = monaco.editor.create(editorEl, {
-      model,
-      minimap: {
-        enabled: false,
-      },
-      scrollBeyondLastLine: false,
-      automaticLayout: true,
-      fontFamily: `ui-monospace, "Fira Code", SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace`,
-      scrollbar: {
-        useShadows: false,
-      },
-      renderLineHighlight: "line",
-      renderLineHighlightOnlyWhenFocus: true,
-      padding: {
-        top: 12,
-        bottom: 12,
-      },
-    });
-
-    const messageContribution = monacoEditor.getContribution("editor.contrib.messageController");
-    readOnlyEditHandler = monacoEditor.onDidAttemptReadOnlyEdit(() => {
-      // eslint-disable-next-line -- MessageController class TS typing is not exported.
-      (messageContribution as any).showMessage("Cannot edit already signed document", monacoEditor.getPosition());
-    });
-
-    const initialVersion = model.getAlternativeVersionId();
-    let currentVersion = initialVersion;
-    let lastVersion = initialVersion;
-
-    goblError.subscribe((goblErr) => {
-      if (!goblErr) {
-        monaco.editor.setModelMarkers(model, "gobl", []);
-        return;
-      }
-
-      monaco.editor.setModelMarkers(model, "gobl", [
-        {
-          message: `${goblErr.message} (code: ${goblErr.code})`,
-          severity: monaco.MarkerSeverity.Error,
-          startLineNumber: 1,
-          startColumn: 1,
-          endLineNumber: 1,
-          endColumn: 1,
-        },
-      ]);
-    });
-
-    unsubscribeEditor = editor.subscribe((value) => {
-      // To keep undo/redo in the editor working, only overwrite the model
-      // contents when the current editor model value isn't the same as the new
-      // store value.
-      if (monacoEditor.getValue() === value) {
-        return;
-      }
-
-      monacoEditor.executeEdits("gobl", [
-        {
-          range: model.getFullModelRange(),
-          text: value,
-          forceMoveMarkers: true,
-        },
-      ]);
-      monacoEditor.pushUndoStop();
-    });
-
-    monacoEditor.onDidChangeModelContent(() => {
-      const value = monacoEditor.getValue();
-      editor.set(value);
-
-      const versionId = model.getAlternativeVersionId();
-      if (versionId < currentVersion) {
-        // Undo occured.
-        redoAvailable.set(true);
-        if (versionId === initialVersion) {
-          // No more undo items.
-          undoAvailable.set(false);
-        }
-      } else {
-        if (versionId <= lastVersion) {
-          // Redo occured.
-          if (versionId == lastVersion) {
-            // Redo of last change occured.
-            redoAvailable.set(false);
-          }
-        } else {
-          // New operation pushed. Disable redo.
-          redoAvailable.set(false);
-          if (currentVersion > lastVersion) {
-            lastVersion = currentVersion;
-          }
-        }
-        undoAvailable.set(true);
-      }
-      currentVersion = versionId;
-
-      validateSchema(value);
-    });
-
-    monaco.editor.onDidChangeMarkers(() => {
-      problems.set(monaco.editor.getModelMarkers({}));
-    });
-
-    monacoEditor.onDidChangeCursorPosition((event) => {
-      lineNumber = event.position.lineNumber;
-      column = event.position.column;
-    });
-
-    document.fonts.ready.then(() => {
-      monaco.editor.remeasureFonts();
-    });
-
-    envelope.subscribe((value) => {
-      monacoEditor.updateOptions({ readOnly: Boolean(value?.sigs) });
-    });
-
-    document.addEventListener("undoButtonClick", handleUndoButtonClick, true);
-    document.addEventListener("redoButtonClick", handleRedoButtonClick, true);
-  });
-
-  onDestroy(() => {
-    unsubscribeEditor();
-    model.dispose();
-    monacoEditor.dispose();
-    readOnlyEditHandler.dispose();
-    document.removeEventListener("undoButtonClick", handleUndoButtonClick, true);
-    document.removeEventListener("redoButtonClick", handleRedoButtonClick, true);
-  });
-
-  function validateSchema(value: string) {
-    if (!jsonSchemaURL) {
-      return;
-    }
-
-    try {
-      const parsed: Record<string, unknown> = JSON.parse(value);
-      if (parsed.$schema !== jsonSchemaURL) {
-        monaco.editor.setModelMarkers(model, "gobl-builder", [
-          {
-            message: `Property "$schema" must be \`${jsonSchemaURL}\`.`,
-            severity: monaco.MarkerSeverity.Error,
-            startLineNumber: 1,
-            startColumn: 1,
-            endLineNumber: 1,
-            endColumn: 1,
-          },
-        ]);
-      } else {
-        monaco.editor.setModelMarkers(model, "gobl-builder", []);
-      }
-    } catch (e) {
-      monaco.editor.setModelMarkers(model, "gobl-builder", []);
-    }
-  }
-
-  function handleUndoButtonClick() {
-    monacoEditor.updateOptions({ readOnly: false });
-    monacoEditor.trigger("undoButton", "undo", null);
-  }
-
-  function handleRedoButtonClick() {
-    monacoEditor.trigger("redoButton", "redo", null);
-  }
-
-  function handleProblemClick(problem: monaco.editor.IMarker) {
-    return function () {
-      monacoEditor.setPosition({ lineNumber: problem.startLineNumber, column: problem.startColumn });
-      monacoEditor.revealLineInCenter(problem.startLineNumber);
-    };
-  }
-
-  function handleDrawerToggle() {
-    drawerClosed = !drawerClosed;
-  }
 </script>
 
 <div class="flex flex-col h-full border-t-2 border-t-slate-200">
-  <div class="flex-1 overflow-hidden" bind:this={editorEl} />
-
-  <div
-    class="flex-none px-4 py-2 bg-zinc-700 text-white text-xs border-b-gray-600 flex items-center gap-6"
-    on:dblclick={handleDrawerToggle}
-  >
-    <div>
-      <span class="mr-1">
-        {#if errorCount > 0}
-          <ErrorIcon />
-        {:else}
-          <SuccessIcon />
-        {/if}
-      </span>
-      <span class="align-middle">
-        {errorCount}
-        {errorCount === 1 ? "error" : "errors"}
-      </span>
+  <div class="relative flex-1 overflow-hidden">
+    <div class="h-full absolute inset-0" class:invisible={$editorViewType !== "code"}>
+      <CodeEditor {jsonSchemaURL} />
     </div>
-    <div class="flex-1">
-      <span class="mr-1">
-        {#if warningCount > 0}
-          <WarningIcon />
-        {:else}
-          <SuccessIcon />
-        {/if}
-      </span>
-      <span class="align-middle">
-        {warningCount}
-        {warningCount === 1 ? "warning" : "warnings"}
-      </span>
+    <div class="h-full absolute inset-0" class:invisible={$editorViewType !== "form"}>
+      <FormEditor />
     </div>
-    <div>Ln {lineNumber}, Col {column}</div>
-    <button class="align-middle" on:click={handleDrawerToggle}>
-      {#if drawerClosed}
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
-          <path
-            fill-rule="evenodd"
-            d="M14.707 12.707a1 1 0 01-1.414 0L10 9.414l-3.293 3.293a1 1 0 01-1.414-1.414l4-4a1 1 0 011.414 0l4 4a1 1 0 010 1.414z"
-            clip-rule="evenodd"
-          />
-        </svg>
-      {:else}
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
-          <path
-            fill-rule="evenodd"
-            d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-            clip-rule="evenodd"
-          />
-        </svg>
-      {/if}
-    </button>
   </div>
-
-  {#if !drawerClosed}
-    <div
-      class="flex-none h-36 py-2 overflow-auto font-mono text-xs text-white bg-zinc-800"
-      transition:slide={{ duration: 300 }}
-    >
-      {#if $editor === ""}
-        <p class="m-4">
-          <span class="mr-2"><LightbulbIcon /></span><span class="align-middle"
-            >Warnings, errors and tips are shown in this area.</span
-          >
-        </p>
-      {/if}
-      {#if $editor !== "" && $problems.length === 0}
-        <p class="m-4">
-          <span class="mr-2"><LightbulbIcon /></span><span class="align-middle"
-            >Use the action buttons in the menu bar.</span
-          >
-        </p>
-      {/if}
-      <ul>
-        {#each sortedProblems as problem}
-          <li class="block cursor-pointer px-4 py-1 hover:bg-zinc-700" on:click={handleProblemClick(problem)}>
-            <EditorProblem {problem} />
-          </li>
-        {/each}
-      </ul>
-    </div>
-  {/if}
 </div>
