+++ src/lib/editor/form/FormEditor.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import NewBlockButtons from "./NewBlockButtons.svelte";
+</script>
+
+<div class="h-full px-6 py-4">
+  <NewBlockButtons />
+</div>
